restore git config checking
It's fast where it needs to be fast (WSL)
And if we set this to hide where the .2s delay would have mattered?
Then there's only the .2s delay, and not the 2s + .2s delay.

To do that:
Bake bash-it.hide-status into %USERHOME%/.gitconfig, and use the
~/bin/git redirection to read it. That way, there's no problem with slow
Windows calls in WSL2