allow typos to be determined based on the current folder (`test -f ./pom.xml`, for instance)
The prior code gets around `-r`'s restricted mode rule "not allowed to
source files with / in their name" by cd'ing there. This time, we use
BASH_ENV to start up bash with those commands run.

I drop the --norc and --noprofile, because this bash is not run as
either an interactive or a login shell, so those don't matter.